updatate double to inpc

+++ src/main/java/mprj/mp/br/calculos/controller/inpcController.java
@@ -60,7 +60,7 @@ public class inpcController {
         List<INPC> lista =   inpcRepository.findByJoinedDateBetweenNative(st,ed);
 
         // Abaixo é igual em todos o controllers
-        float Valor3 = 0 ;
+        double Valor3 = 0 ;
         float Valor4=  0;
         double valorJuros = 0.0;
         JSONArray jsonArray = new JSONArray();
